start index as usize

+++ src/scripts.rs
@@ -225,7 +225,7 @@ pub fn cli_entrypoint(py: Python) -> Result<()> {
             let slice = view.slice_axis(
                 Axis(0),
                 Slice::new(
-                    pano_args.start.unwrap_or(0),
+                    pano_args.start.unwrap_or(0) as isize,
                     pano_args.end.map(|v| v.min(view.len_of(Axis(0)) as isize)),
                     1,
                 ),
